Add Range class and corresponding test cases
In this commit, a new Range class under the 'Bizbozo\Adventofcode2023\Ranges' namespace has been created with methods for creating a range, checking if a number belongs to the range, calculating intersection and difference of two ranges, and shifting the range. Additionally, corresponding test cases have been added in the RangeTest class to validate all these methods of the Range class. Exceptions for invalid range have also been implemented.