Add notes and blog for day 05 in readme.md
Added a new section "Notes" in the readme.md file and linked a new blog post about the revisit to the day 05 problem. The blog post provides an insight on how the issue with range calculations was resolved using a new Range class resulting in better readability and maintainability of the code. It also details usage of Jetbrains AI in solving the problem.